requirements.txt
pystray
plyer
pillow
